Speed up maxctrl_basic
The test now batches the commands into a file and uploads it into the
VM. Feeding the commands into maxctrl there avoids the overhead of
starting up Node.js which starts to be visible when commands are
repeatedly executed.

This commit speeds up the test from a total runtime of about 37 seconds
down to 17 seconds when run locally.