MXS-5344: Don't create constraint on _id
The constraint was too restricting on what types of JSON data can be
inserted into MariaDB. Especially when combined with kafkacdc, this
limit made it both hard to debug the insertion failures and also
required a manual step to make it work that didn't bring any practical
benefits for the original use-case.

As multiple NULL values are allowed in a UNIQUE KEY, the uniqueness of
the _id field is still guaranteed but the presence of it is no longer
required.