Update installation instructions
The links in the instructions were old and needed updating.

The use of MariaDB repositories wasn't documented as the primary method
of installing MaxScale when it should have given that it is much easier
than having to download packages manually.

Replaced the use of `dpkg -i` with the more modern use of `apt` which
nowadays can handle installing a package from the command line while
also installing dependencies at the same time.