MXS-5355: Discard pending retries on trx replay start
There was a false assumption in readwritesplit that a transaction replay
would only get started if a failure would occur. In reality it's also
possible for a transaction replay to start if the current target of the
transaction is no longer a valid routing target or it responds with an
error that should trigger a transaction replay. It's also possible for
the aforementioned event to happen right before a network disconnection
happens which triggers yet another transaction replay to take
place. This means that it's possible (and valid) that two transaction
replays are started before the delay_routing() call of the first one
resolves. In this case, m_pending_retries count is positive when
start_trx_replay() gets called.

Previously, when two transaction replays were started in short
succession, the transaction replay was simply started again and the
first statement of the transaction was queued for execution. However,
this caused one statement from the previous replay attempt to be
executed as a part of the current replay attempt, potentially causing a
statement to be executed out-of-transaction.

The solutions to this are to either not start a transaction replay if
one is already ongoing or to discard the pending delay_routing()
calls. Checking that only one transaction replay is ongoing is currently
more complex than it would be to simply discard the queries that are
queued with delay_routing(). The latter can be implemented with an extra
counter that, when positive, causes the queries to be discarded instead
of executed.

The order in which the delay_routing() calls resolve are not guaranteed
to be the same as the order in which they are called. This happens as
the delay may cause a later delay_routing() call to be executed before
an older one. By moving the queuing of the retries into readwritesplit
itself, the order in which the queries get executed remains the same as
the order in which the retry_query() functions are called. This means
that eventually, the correct query gets executed once all of the
delay_routing() calls have resolved. As the first query of a retried
transaction is always executed with a one second delay, the actual time
that it takes to execute the first query does not change.