MXS-5359: Always restart trx replay on routing failure
If the routing of the query fails and a transaction replay is ongoing,
the replay should always be restarted from the beginning. This way, the
queries of the retried transaction only get queued either from
start_trx_replay() or trx_replay_next_stmt(), never from the code in
routeQuery() itself that usually deals with queries replayed due to
delayed_retry. With this commit, the "deadlock" that would occur during
a switchover is avoided and the overall logic of the transaction replay
ends up being more straightforward.