MXS-5363: Write GDB output to a temporary file
The problem with using popen() to communicate with GDB is that it uses a
pipe and the pipe has a limited size before GDB's write gets
blocked. The pipe is also never drained as MaxScale is stopped by GDB
while it writes into the pipe. By redirecting GDB's output into a
temporary file, this problem goes away.