MXS-5356: Prevent reconnections if GET_LOCK() is used
The use of GET_LOCK() with readwritesplit was not safe if
master_reconnection was enabled: if GET_LOCK() was used, it could be
lost without the client being aware of it if a reconnection happened.

Now the use of GET_LOCK() will disable reconnections in
readwritesplit. Currently the detection is relatively simple and the
unlocking of a lock is not detected and thus the "unsafeness" is never
cleared away.

Added a test case that reproduces the problem.