List install byproducts of external projects
This'll make CMake aware of the byproducts directly which helps it work
with the Ninja build system. The benefits of this is that Ninja by
default runs everything in parallel (with a reasonable default) which
speeds up builds of the external projects. It also shows if there are
any implicit dependencies between targets that happen to be built in the
right order due to how makefiles order the build.

For the hiredis build, as it has had some problems in the past, the
configuration for it forces the build system to use Unix makefiles.