Inject version function into Java based tests
The maxscale_version() function will return the MaxScale version as a
number that the JDBC and R2DBC tests can use to check whether a
particular test can be run or not. Once the pull requests on the
connector test suites go through, this commit should fix the connector_j
and connector_r2dbc tests.