Adapt JSON Path code to old mxb::strtok
The version of mxb::strtok in 22.08 is different in that it takes and
returns a std::string. The code needed a few changes to make it work.

The lack of mxb::cat also needed changes. The code now uses MAKE_STR
which is a bit less efficient but it should be adequate.