MXS-5438: Update GTIDs in galeramon
Galeramon now updates the GTID positions of the servers. This can then
be used by readwritesplit to add support for the fast causal_reads
modes.

Added a note to the readwritesplit limitations about the fast modes.

Added a derived version of the test case for causal_reads=fast and
causal_reads=fast_global that uses Galera as the backend.