MXS-5404 Discard monitor journal if different config file
The configuration file path is now stored in the monitor journal.
That makes it possible to detect if the journal was created for
a configuration file, different from the one currently used. If
so, the journal is discarded.

This is not foolproof since dynamic changes are stored in the
same directory, regardless of the static configuration file used.
So there may still be changes that should cause the journal to be
discarded.