MXS-5432: Never discard queued packets
This is not safe to do as the upper layer still expects responses for
the commands, even if they are going to be ignored. A potential solution
to this kind of a situation would be to replace the COM_STMT_PREPARE
with a COM_QUERY that generates a response but does nothing, for example
DO 1, but this would be more complex than the current approach of just
discarding connections that lag behind too much.

Added a test case that reproduces the problem and verifies that removing
the packet discarding fixes the problem.