MXS-5443: Add test case
Added a test case that reproduces the original problem with 21.06.18 and
confirms that commit 56a205538636f2de72f41cf9b16bfffb8b453728 would've
fixed it.

Also added a test case for the bug that was fixed by moving the delay
queue blocking detection into readwritesplit where the role of the
backend was known. This also confirms that the previous fix wouldn't
have been adequate and would've resulted in problems where lots of small
packets are expected to get queued up.