MXS-5463 Fix log filter by timestamp
Fixed an issue where the latest log was not fetched when re-entering
the GUI Logs Archive page for the second time. Instead, it displayed
logs from the previous timestamp, with the latest log only appearing
after a manual page refresh (F5). This occurred because the `logFilters`
getter was cached, and the `date_range` value was either a string or
constant (`NOW`), causing the getter to return the old timestamp.

This fixed it by turning the getters to get date range and log filter
params into a function.