MXS-5480: Fix disable_sescmd_history=true
The history was assumed to be non-empty when the pruning
happened. Previously this was true but in the current 23.08 branch, this
is not true anymore as the history removal ends up deleting the GWBUF
that was just inserted into it when disable_sescmd_history=true is used.

Added a test case that reproduces the problem and verifies that the fix
works.