MXS-5484 Not an error if a requested gtid is not found
This was a problem in 24.02 because the gtid search in binlogs
no longer searches the entire file, so new domains in a file
would not be found.

This also prepares for changing primary sync to be made faster
in the case where the binlog directory is wiped out (allowed
when maxscale is not running).