MXS-5492: Make idle_session_pool_time=0s fairer
The dcall that is added when idle_session_pool_time is configured to be
below one second cause any failure to pool a connection to then not pool
connections for one second. This caused severe monopolization of the
connections when idle_session_pool_time=0s was used. By never using the
dcall for idle_session_pool_time=0s, the pooling is always attempted
whenever a request completes and the client goes idle. This appears to
work a lot better in the special case where idle_session_pool_time=0s is
used.

To cover the case where the pooling fails right as the client goes idle,
an additional check is done in the tick() function of the session to see
if connections can be pooled.

Another part of the unfairness of idle_session_pool_time=0s was caused
by the fact that the waking up of idle connections was done with an
execute() call with a queued event. This made it so that the waking up
of clients was interleaved with new requests from recently active or new
clients. This allowed some clients to skip ahead of the queue and in
practice this seemed to be the recently active connections which led to
the monopolization of the connections by a handful of sessions.

By using a lcall() instead of a queued worker message, the waking up of
idle connections happens before epoll_wait() returns new events. This
allows the waiting sessions to grab the free connections before the
recently active clients have a chance to steal them.