Add debug endpoint for mxs2057_systemd_watchdog
Previously, the test relied on the inefficient execution of certain
regular expression patterns that caused a thread to stall for long
times. This has proven to be very unstable and inconsistent. Adding a
separate debug endpoint for testing the watchdog functionality will make
it possible reliably test it.

The new maxscale/debug/hang endpoint is only available in debug
builds. This is to prevent it from being accidentally called in
published release mode packages. If the mxs2057_systemd_watchdog test is
run with a release build, the test detects that the call failed and it
skips the watchdog test.

Also increased the wait time to twice of the watchdog interval. This
should avoid any false positives due to timing differences.