MXS-5520: Only mask passwords in ParamPassword
The act of decrypting them should be left to the code that uses the
configuration parameter. This way the encryption keys only need to be
loaded when the password is actually used which fixes the problem
described in MXS-5520.

This also has the desired side-effect of retaining the original
encrypted password whenever it is persisted. Previously, when a user
configured an encrypted password, it would be decrypted at the point
where it was configured and then later encrypted again when it needed to
be stored on disk. This resulted in the ciphertext of the password
changing which made it difficult to verify exactly which password was in
use.