MXS-5522: Ignore changes to port, address and socket
When configuration synchronization is updating a listener, the port,
address and socket of it should be ignored as they're not configurable
at runtime and must be done by deleting the of the object and creating
it again with the new configuration. The changes to the configuration
validation now report it as errors so special handling for them is
required in the config manager code.