MXS-5527: Always look for last_gtid variable
The type of the result doesn't matter now that resultsets can be
associated with a transaction due to INSERT...RETURNING.

The documentation now also mentions the connector type as a limitation
when INSERT...RETURNING is being used.

Added a test case that reproduces the problem and verifies that the fix
works. Unfortunately, because the Connector/C does not support the
DEPRECATE_EOF protocol, the test has to use Connector/Node.js.