MXS-5522: Fix config sync checksums
The modifications to the listener JSON were not reverted after the
listener was configured. This caused the checksum to differ from the
source MaxScale's checksum.

The checksum calculation also used the default to_string() method from
mxb::Json which did not sort or compact the input. Relying on the
default behavior of it for the checksum is not good and using
mxb::json_dump() directly with the desired flags is better.