MXS-5533: Set log scope during configuration
The scope in which any warnings or errors are logged by generic code is
then much clearer. For example, without this log scope the use of the
session_trace parameter in services would log a warning about
session_trace being deprecated when in fact it's only deprecated when
used in services.