MXS-5505 Improve gtid search
Currently, once a replica connects, the gtids from the
"change master" cmd are searched for and the file(s) are
read from the erliest point of those gtids. This is fine
for normal replication, but in multidomain replication
some gtids could be far back in terms of files and both
the search and subsequent file event fetch scan
files unnecessarily.

In this commit the search finds the next gtids in sequence,
i.e. instead of "1-1-10, 2-2-1000000", the search is for
"1-1-11, 2-2-1000001". If the last gtids for each domain
in the binlogs are what the replica has (1-1-10, 2-2-1000000),
the search only scans the last file and the read position
for fetch is set to the end of that file.