Test connection liveness with a query
The isConnectionOk() function, which wrongly returned false if the
connection is okay, was not a reliable way to test the liveness of a
connection as errors generated by MaxScale might've not propagated all
the way to the connector. A more reliable way to test it is to execute a
SQL query and check that it gets rejected.