MXS-5525 Reduce false positives by adding the table if known
A statement like "SELECT * FROM t1 WHERE f = 1" will be reported
to refer to a field "f", i.e. only the column name without table
information. If there is masking rule for "t2.f" ("t2" not "t1"),
it will match, because if the table is not provided the assumption
must be that the table might be "t2".

However, if the statement only refers to one table ("t1" above),
we can patch the field info with that information with the result
that the rule no longer matches.

If there are more tables than one, nothing can be done as, without
access to the schema, we do not know which table the field refers
to.