MXS-5541: Fix REST-API filtering in 24.02
If a request that produced an error also used a filter expression, the
lack of a valid response body was treated as a filter expression error
which overwrote the error object. By only filtering successful responses
and checking the body type before parsing the expression, the correct
error is returned.