MXS-5544: Add hints to all PS commands
This only affects the schemarouter which follows routing hints
regardless of the query type. The COM_STMT_CLOSE commands also did not
receive a routing hint as they were assumed to be routed to all
backends. For readwritesplit, all binary protocol command that must be
routed to all backends ignore routing hints.

Added a test case that reproduces the original problem and verifies that
the fix works.