MXS-5588: Fix GTID parsing and probing
The GTID parsing assumed a non-empty GTID value which was true whenever
it was generated by the server via the session tracking
mechanism. However for the causal_reads=universal mode, the GTID is
queried using `SELECT @@gtid_current_pos` which may return an empty
string in the case of Galera.

The GTID parsing now uses the <charconv> functions to both parse and
serialize the values. This results in the correct behavior when
malformed inputs are provided. Additionally, this gets rid of all memory
allocations that previously took place during the GTID parsing.

Also switched the GTID probe query to use `SELECT
@@gtid_binlog_pos`. This results in the correct expected behavior where
the last binlogged GTID on the server is used as the sync point instead
of potentially the last replicated GTID. This also make the
causal_reads=fast_universal mode work correctly with Galera clusters.