MXS-5590: Fix connection reuse in microhttpd
The MHD_AccessHandlerCallback handler is called multiple times. For POST
uploads with a body, it is called several times. For GET requests, it's
also called multiple times, first after processing the headers and a
second time after the request has been fully read. If a response was
generated already in the first callback, the rest of the response would
be discarded even if the request had no body. This is what resulted in
the "Connection: close" header being automatically added to all requests.

By returning early the first time the callback is called, the behavior
is correct and microhttpd successfully reuses the TCP connection. This
also revealed that a lot of the code was actually dead code and was
never called.