Fix schemarouter_duplicate
The test assumed that the servers had lost the Slave state when the
connection attempt was made. If the status had not yet been lost and it
was lost during the updating of the shard maps, the connections would
get closed. This would then cause the test to fail as the duplicate
tables would not be detected since the connections were closed before
the mapping was complete.