MXS-5588: Fix debug assertion in readwritesplit
The SQL that the GTID probe used was also used in a debug
assertion. Since the variable name was corrected in the previous commit
to fetch the binlogged position, the assertion started to fail.

The documentation was also out of date and did not mention the change in
the variable.