MXS-5582: Add test case
Added a convenience macro for declaring test cases which will be needed
to fix the test case numbering when merging tests upwards. Also made the
test easier to run manually by resetting the state at the start of the
test which allows the test to be interrupted without causing the next
run to immediately fail.

The test case itself is simple: create a service that uses the existing
monitor and its servers as the `cluster`. The second MaxScale will fail
to apply due to the incorrect order in which relationships are
processed.