MXS-5582: Use relationships correctly in config sync
The object relationships are expressed as bidirectional dependencies in
the REST-API JSON output which is intended to be usable to "navigate"
through the API. The same JSON cannot always be used to define the
object itself as there are restrictions on what kinds of relationships
an object can have and in which order the relationships go.

For example, when a service uses a cluster (a monitor) which in turn
uses servers, the servers will have a relationship link to each service
that uses it. This is useful when displaying information but the
relationship is not a "true" one in the sense that it doesn't correctly
express the way in which that relationship is formed. The correct way to
express the relationships would be to not have the service relationship
whenever the service uses a cluster but this makes the information
retrieval more complex. Given that the configuration synchronization is
an internal process which knows the layout of the JSON, it can process
the relationships correctly without having to have the REST-API express
the relationships in a strict manner.

Servers, which can never depend on anything, should not have any
relationships when being processed via configuration
synchronization. Similarly, the monitors should only have the servers
relationship while it is being processed. When the services are
processed, they'll then correctly choose between using targets or the
cluster.