Run cache storage test cases in parallel
The test_cache_lru_inmemory test takes about 30 seconds to run and the
test_cache_storage_inmemory test takes about 17 seconds. Running the
individual test cases in parallel cuts the testing time of both to under
10 seconds.

The tests can be run in parallel as each test case allocates a separate
Storage instance.

Since the interleaved output of the parallel tests is hard to read, the
NO_PARALLEL=1 environment variable can be used to run tests
serially. This is useful when debugging test failures.