MXS-5599: Fix conditional header processing
The conditional headers are now processed before the result is generated
which removes the side-effects that their execution previously
caused.

By using a version counter that gets incremented at the same time the
modification timestamp is updated, the Etag generation can be done
without the contents of the object itself. This removes the duplicated
JSON dump that previously was used to generate the Etag value. This
change works because the only way to modify an object is via the
REST-API and all methods that do modify it (POST, PATCH, DELETE) update
the timestamps and the version counter.

The conditional execution headers are intended to be used to prevent
concurrent external modifications of objects, not to detect when the
internal state of the object has changed.