Fix debug assertion in monitor creation
The mxs118 test hit a debug assertion where the add_server() call in
Monitor::post_configure() was always assumed to work. This was not the
case when the servers were already monitored by another monitor in
MaxScale.

Also improved the error message to state the monitor to which the
addition was attempted.