Parallelize test_worker
The test now runs all of the test cases in parallel. The move testing
also always moves instantly instead of waiting for 10 miliseconds. To
offset this, the test is now run 10 times in parallel in the hopes that
it'll increase the chance of finding any concurrency related bugs.

Combined with the reduction in test timings done in 21.06, the test time
goes down from 17 seconds to about two seconds.