MXS-5607: Clarify readwritesplit routing behavior
The tiebreaker behavior in least_current_operations was not
documented. Similarly, the monitor's measurement of the ping time was
not documented for adaptive_routing.

Also added a paragraph about how the connections and queries are
measured and that connections and queries done outside of MaxScale are
not taken into account.