Generate test data in parallel in test_compression
Profiling showed that most of the time went into generating the random
numbers for the 100MiB of test data that the test used. The compression
and decompression itself was a fraction of the total test time. The data
generation was done serially and parallelizing it cuts the testing time
in roughly half.

Ideally the generation would be done using the execution policy overload
of std::generate but that requires linking against tbb and it's not
available on all of the supported systems which is why a simpler
std::async based approach is used.