Fix config_sync
One of the test cases was actually two separate test cases. The previous
commit removed the reset() call from the wrong function which caused
stale state information to be left.

Splitting the test cases into individual test cases fixes this and also
improves the accuracy of the ouput in case the test fails.