MXS-5608: Add missing ack_write() to rollback
The transaction rollback didn't call ack_write() on the backend which
caused it to leave an extra unacknowledged response on the backend. The
next time that a statement would be routed there, the expected response
queue would get shifted by one with a extra response being returned if
the next command was a session command.