MXS-5606: Read maxctrl credentials from environment
This is a secure way to pass the credentials to maxctrl without having
to store them in a configuration file or give them as command line
arguments.

The implementation of the .env() feature in Yargs combined with
.strict() will cause all unknown environment variables that start with
MAXCTRL_ to be interpreted as options which causes an error if they're
not an option. To limit the likelihood of this happening, only the
MAXCTRL_USER and MAXCTRL_PASSWORD variables are supported.

Added a test case that checks that the reading of the credentials is
possible and that wrong credentials cause an error even if the valid
default credentials are still in place.