Add LOG_OUTPUT_ON_FAILURE to ExternalProject calls
This way the output is logged if an error occurs but it's written to a
file on disk if there are none. Turning off the logging of the output to
a file would make it possible to find out why it fails but in the normal
case where there are no problems, the output would be way too verbose.