MXS-5685: Backport documentation docker changes
A local docker image is a lot more flexible compared to a released one
as it must anyways be done before a release takes place. This also
removes the need to do a post-release maintenance task and allows
development to be completed once the feature itself is complete and
documented.