MXS-5882 Store static config state also for servers
At startup, the initial config from the static config file(s)
should be stored, so that it can be detected whether several
runtime changes restores the state to the initial one. This,
in order to be able to know when to delete a runtime config
file. This was made for all other objects but for servers.