MXS-5769: Store timestamps of watchdog notifications
If a systemd watchdog timeout occurs, the fatal signal handler will show
how many seconds ago the last notification was done. While this is
helpful for detecting actual watchdog timeouts, it doesn't help if the
watchdog timeout is somehow related to the system itself. By storing a
fixed amount of local timestamps, we'll see if there's drift between
them as well as the absolute time at which the notifications were
done. With this information, it should be possible to conclusively state
when the notifications have been sent.