Add dump-network-traffic debug flag
The debug flag now controls whether the network traffic dump is done or
not. When both debug=dump-network-traffic and log_info=true are defined,
the data is hexdumped on the info level. Previously, this data was only
logged in debug builds when log_debug was enabled.

The network data dump is the most efficient way to debug protocol level
issues as it allows the raw network data to be inspected. For setups
that use TLS, a normal network dump won't work as the traffic is
encrypted and decryping it usually would require access to the private
keys. Additionally, the installation of debug package is a lot more work
compared to enabling a debug flag in a release mode build. The debug
packages are also not an option in production setups since they're not
optimized and use ASAN and UBSAN.