MXS-5897: Add test case
The test reproduces the problem and verifies that the fix works and that
the history does not grow if a COM_STMT_PREPARE fails.

Also added one missing mysql_stmt_free() call in the test and moved the
"main test" into a separate sanity_check() function. With the increasing
number of sub-tests, the ability to run one of them becomes easier if
they're all in their own functions.