MXS-5952 When configuring replication in server 11.4, always set MASTER_SSL
MariaDB Server 11.4.1 and later support zero-configuration SSL and
have it enabled by default. This also applies to replication,
meaning that if MASTER_SSL is not set, it is assumed on. This
causes replication to fail if the primary server has SSL disabled.

Fix this by always setting MASTER_SSL when configuring replication
with recent server versions.