Install and use SUSE development tools
A newer compiler than the default one is needed for <filesystem>.
The CC/CXX settings cannot be written to ~/.bashrc, because that
will not necessarily be read when a script is executed.

The file can also not be written to ~, because install_build_deps.sh
is run as root, while build_package.sh is run as whatever account is
used for building.