MXS-5955 Delay assignment of server role status bits during startup
The SERVER_MASTER and SERVER_SLAVE-bits in the journal could be
wrong and cause a write to the wrong server if topology changed
during downtime. Avoid this by only settings SERVER_MAINT and
SERVER_DRAINING bits when reading the journal. The rest of the
bits are set during the first monitor tick.