Use newer Node.js in tests
The Connector/Node.js tests need a newer version of Node.js to run. The
version 16 that was previously used wasn't new enough to make it work.

The test dependency installation also now passes the CMake and Node.js
versions as arguments to the normal dependency installation. This way
CMake and Node.js are only downloaded once.