Fix GCC 15 build failures
The <cstdbool> header has been deprecated since bool is a part of C++.

The SQLite function definitions didn't match their declarations which is
an error with GCC 15.

The avro-c library no longer compiles with GCC 15 as the default
standards have changed to versions that it isn't compatible
with. Explicitly defining the standards solves the problem.