Add missing SoftError to administration.hh
Instead of throwing a `SoftError`, the value
`error::CANNOT_CREATE_INDEX` was thrown instead. This is because the
`(ss.str(), error::CANNOT_CREATE_INDEX)` is the comma operator which
discards the first value and returns the second one. With GCC 15, the
`str()` method of `stringstream` now has `nodiscard` which revealed this
problem.