Fix syncing of maintenance and draining bits
With configuration synchronization, there will be a maintenance or a
draining status change request right when the monitor starts up while
the journaled status was stored in m_status_from_journal. This had the
effect of canceling out the status change request which caused the
config_sync test to fail.

The status change request handling should only be executed if it's not
the first tick. This makes sure that the status from the journal is read
before any requests are processed.